+++ tests/test_env.py
@@ -27,7 +27,7 @@ def test_find_all_from_json_string(env: JSONPathEnvironment) -> None:
 def test_find_all_with_extra_filter_context(env: JSONPathEnvironment) -> None:
     """Test that we can pass extra filter context to findall."""
     rv = env.findall(
-        "$[?(@.some == #.other)]",
+        "$[?(@.some == _.other)]",
         {"foo": {"some": 1, "thing": 2}},
         filter_context={"other": 1},
     )
@@ -49,7 +49,7 @@ def test_find_iter_from_json_string(env: JSONPathEnvironment) -> None:
 def test_find_iter_with_extra_filter_context(env: JSONPathEnvironment) -> None:
     """Test that we can pass extra filter context to finditer."""
     matches = env.finditer(
-        "$[?(@.some == #.other)]",
+        "$[?(@.some == _.other)]",
         {"foo": {"some": 1, "thing": 2}},
         filter_context={"other": 1},
     )
@@ -79,7 +79,7 @@ def test_find_all_async_with_extra_filter_context(env: JSONPathEnvironment) -> N
 
     async def coro() -> List[object]:
         return await env.findall_async(
-            "$[?(@.some == #.other)]",
+            "$[?(@.some == _.other)]",
             {"foo": {"some": 1, "thing": 2}},
             filter_context={"other": 1},
         )
@@ -112,7 +112,7 @@ def test_find_iter_async_with_extra_filter_context(env: JSONPathEnvironment) -> 
 
     async def coro() -> List[object]:
         matches = await env.finditer_async(
-            "$[?(@.some == #.other)]",
+            "$[?(@.some == _.other)]",
             {"foo": {"some": 1, "thing": 2}},
             filter_context={"other": 1},
         )